assert_is_debugged {assertive.code}	R Documentation

Is the input function being debugged?

Description

Checks to see if the input DLL (a.k.a. shared object) is loaded.

Usage

assert_is_debugged(x, severity = getOption("assertive.severity", "stop"))

is_debugged(x, .xname = get_name_in_parent(x))

Arguments

x
Input to check.

severity
How severe should the consequences of the assertion be? Either "stop", "warning", "message", or "none".

.xname
Not intended to be used directly.

Value

is_debugged wraps isdebugged, providing more information on failure. assert_is_debugged returns nothing but throws an error if is_debugged returns FALSE.

See Also

isdebugged.

[Package assertive.code version 0.0-3 Index]