is_error_free {assertive.code}	R Documentation

Does the code run without throwing an error?

Description

Call the code inside a try block and report if an error was thrown.

Usage

is_error_free(x)

Arguments

x
Code to check.

Value

TRUE if the code runs without throwing an error. The result of running the code is contained in an attribute named "result".

Note

Note that this has the side effect of running the code contained in x.

[Package assertive.code version 0.0-3 Index]