assert_all_are_existing {assertive.code}	R Documentation

Does the variable exist?

Description

Checks to see if the input variables exist.

Usage

assert_all_are_existing(x, envir = parent.frame(), inherits = TRUE,
  severity = getOption("assertive.severity", "stop"))

assert_any_are_existing(x, envir = parent.frame(), inherits = TRUE,
  severity = getOption("assertive.severity", "stop"))

is_existing(x, envir = parent.frame(), inherits = TRUE,
  .xname = get_name_in_parent(x))

Arguments

x
Input to check.

envir
Passed to exists.

inherits
Passed to exists.

severity
How severe should the consequences of the assertion be? Either "stop", "warning", "message", or "none".

.xname
Not intended to be used directly.

Value

is_existing is a vectorized wrapper to exists, providing more information on failure (and with a simplified interface). The assert_* functions return nothing but throw an error if is_existing returns FALSE.

See Also

exists.

Examples

e <- new.env()
e$x <- 1
e$y <- 2
assert_all_are_existing(c("x", "y"), envir = e)
#These examples should fail.
assertive.base::dont_stop(assert_all_are_existing(c("x", "z"), envir = e))

[Package assertive.code version 0.0-3 Index]