assert_is_valid_r_code {assertive.code}	R Documentation

Is the input valid R code?

Description

Check to see if the input is a valid (parseable) R code.

Usage

assert_is_valid_r_code(x, severity = getOption("assertive.severity",
  "stop"))

is_valid_r_code(x, .xname = get_name_in_parent(x))

Arguments

x
Input to check.

severity
How severe should the consequences of the assertion be? Either "stop", "warning", "message", or "none".

.xname
Not intended to be used directly.

Value

TRUE if the input string is valid R code.

See Also

parse

Examples

is_valid_r_code("x <- 1 + sqrt(pi)")
is_valid_r_code("x <- ")
is_valid_r_code("<- 1 + sqrt(pi)")

[Package assertive.code version 0.0-3 Index]